+++ front/src/pages/trade/chart/period/PeriodBackground.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
-import { IChartItem, Theme } from '@src/types';
-import userAtom, { IUser } from '@recoil/user';
+import { IChartItem, Theme, IUser } from '@src/types';
+import userAtom from '@recoil/user';
 import {
 	TChartType,
 	MAKE_CLEAR_OFFSET,
