+++ front/src/recoil/user/index.ts
@@ -1,6 +1,1 @@
-import atom from './atom';
-import type { IUser } from './atom';
-
-export { IUser };
-
-export default atom;
+export { default } from './atom';
